,

Fixation positions of the eye-tracking data of Frank, Fernandez Monsalve, Thompson et al. (2013)

Fixation positions of the eye-tracking data of Frank, Fernandez Monsalve, Thompson et al. (2013).

Usage

data_frank2013_et_fix

Format

A data frame:

subj: Subject id.

sent_id: Sentence unique id.

gaze_x: Horizontal pixel coordinate of fixation location (0 = left edge; 39 = left margin; 1024 = right edge). Each letter (including space and punctuation) was 14 pixels wide.

gaze_y: Vertical pixel coordinate of fixation location (0 = top of display; 768 = bottom of display).

fix_duration: Duration of the fixation in ms.

letter_n: Fixated letter position in sentence (including spaces), even if the fixation was to the left of the left margin or to the right of the last letter. Position 0 is directly to the left of the first letter.

word_n: Position of the word in the sentence.

word: Presented word.

blink: Indicates whether a blink was detected directly 'before' or 'after' the current fixation (or 'both' before and after).

Details

As in the original dataset: "Within each sentence, fixations are listed in the order in which they occurred. If any fixation during sentence presentation was registered as being outside the display bounds, this was regarded as a tracking error and no data is presented for that sentence. The letter and word position are NaN if gaze_y is outside the range (300,475) because that was considered too far above or below the presented sentence."

References

1 S. L. Frank, I. Fernandez Monsalve, R. L. Thompson, et al. ``Reading Time Data for Evaluating Broad-Coverage Models of English Sentence Processing''. In: Behavior Research Methods 45.4 (dec. 01, 2013), pp. 1182--1190. ISSN: 1554-3528. DOI: 10.3758/s13428-012-0313-y.

See also