,

Cleaned self-paced reading data of Frank, Fernandez Monsalve, Thompson et al. (2013)

A cleaned version of dataset "data_frank2013_spr_complete": No sentences with typos, only native speakers, and subjects with accuracy over 0.8 from Frank, Fernandez Monsalve, Thompson et al. (2013)

Usage

data_frank2013_spr

Format

An object of class tidytable (inherits from data.table, data.frame) with 165040 rows and 12 columns.

References

1 S. L. Frank, I. Fernandez Monsalve, R. L. Thompson, et al. ``Reading Time Data for Evaluating Broad-Coverage Models of English Sentence Processing''. In: Behavior Research Methods 45.4 (dec. 01, 2013), pp. 1182--1190. ISSN: 1554-3528. DOI: 10.3758/s13428-012-0313-y.

See also