,

Self-paced reading data of Frank, Fernandez Monsalve, Thompson et al. (2013)

A dataset containing the self-paced reading data of Frank, Fernandez Monsalve, Thompson et al. (2013). Notice that after collecting the self-paced-reading data, three typos were found:

sentence 43 : "Scott" was "Sott"

sentence 269 : "at" was "that"

sentence 337 : "Margaret" was "Margeret"

Usage

data_frank2013_spr_complete

Format

A data frame with rows and variables:

subj: Subject id.

sent_id: Sentence unique id.

sent_n: Position of sentence in the presentation sequence.

answer_time: Time in ms between question presentation and response, or NaN if there was no question.

word_n: Position of the word in the sentence.

word: Presented word.

RT: Time in ms between word presentation and key press.

acc_comprehension: Correctness of response to comprehension question.

age: Subject's age in years.

age_en: Age at which subject began learning English (0 for native speakers).

sex: Subject's sex (f/m).

hand: Subject's handedness (r/l).

correct_perc: Fraction of correct responses to comprehension questions.

typo: Whether there is a typo in the word.

References

1 S. L. Frank, I. Fernandez Monsalve, R. L. Thompson, et al. ``Reading Time Data for Evaluating Broad-Coverage Models of English Sentence Processing''. In: Behavior Research Methods 45.4 (dec. 01, 2013), pp. 1182--1190. ISSN: 1554-3528. DOI: 10.3758/s13428-012-0313-y.

See also