,

Get a matrix with the log probabilities of possible word given its previous context using a causal transformer

Get a matrix with the log probabilities of possible word given its previous context using a causal transformer model from Hugging Face.

Usage

get_causal_log_prob_mat(
  x,
  .by = rep(1, length(x)),
  model = "gpt2",
  add_bos_token = NULL,
  stride = 1,
  config_model = NULL,
  config_tokenizer = NULL
)

Arguments

x

Vector of words, phrases or texts.

.by

Vector that indicates how the text should be split.

model

Name of a pretrained model stored on the huggingface.co. (Maybe a path to a model (.pt or .bin file) stored locally will work.)

add_bos_token

Whether to include beginning of text special tokens. By default acts as the AutoTokenizer.

config_model

List with other arguments that control how the model from hugging face is accessed.

config_tokenizer

List with arguments that control how the tokenizer from hugging face is accessed.

Value

A matrix.

Details

For more about causal models, see (https://huggingface.co/course/chapter7/6). Using the config_model and config_tokenizer arguments, it's possible to control how the model and tokenizer from hugging face is accessed, see from_pretrained for details. In case of errors check the status of https://status.huggingface.co/