,

Deprecated functions in package pangoling.

The functions listed below are deprecated and will be defunct in the near future. When possible, alternative functions with similar functionality are also mentioned. Help pages for deprecated functions are available at help("<function>-deprecated").

Usage

masked_tokens_tbl(
  masked_sentences,
  model = getOption("pangoling.masked.default"),
  add_special_tokens = NULL,
  config_model = NULL,
  config_tokenizer = NULL
)

masked_lp(
  l_contexts,
  targets,
  r_contexts,
  ignore_regex = "",
  model = getOption("pangoling.masked.default"),
  add_special_tokens = NULL,
  config_model = NULL,
  config_tokenizer = NULL
)

causal_next_tokens_tbl(
  context,
  model = getOption("pangoling.causal.default"),
  checkpoint = NULL,
  add_special_tokens = NULL,
  config_model = NULL,
  config_tokenizer = NULL
)

causal_lp(
  x,
  by = rep(1, length(x)),
  l_contexts = NULL,
  ignore_regex = "",
  model = getOption("pangoling.causal.default"),
  checkpoint = NULL,
  add_special_tokens = NULL,
  config_model = NULL,
  config_tokenizer = NULL,
  batch_size = 1,
  ...
)

causal_tokens_lp_tbl(
  texts,
  model = getOption("pangoling.causal.default"),
  checkpoint = NULL,
  add_special_tokens = NULL,
  config_model = NULL,
  config_tokenizer = NULL,
  batch_size = 1,
  .id = NULL
)

causal_lp_mats(
  x,
  by = rep(1, length(x)),
  sorted = FALSE,
  model = getOption("pangoling.causal.default"),
  checkpoint = NULL,
  add_special_tokens = NULL,
  config_model = NULL,
  config_tokenizer = NULL,
  batch_size = 1,
  ...
)

masked_tokens_tbl

For masked_tokens_tbl, use masked_tokens_pred_tbl.

masked_lp

For masked_lp, use masked_targets_pred.

causal_next_tokens_tbl

For causal_next_tokens_tbl, use causal_next_tokens_pred_tbl.

causal_lp

For causal_lp, use causal_targets_pred or causal_words_pred.

causal_tokens_lp_tbl

For causal_tokens_lp_tbl, use causal_tokens_pred_lst.

causal_lp_mats

For causal_lp_mats, use causal_pred_mats.