,

Returns the configuration of a causal model

Returns the configuration of a causal model

Usage

causal_config(
  model = getOption("pangoling.causal.default"),
  config_model = NULL
)

Arguments

model

Name of a pre-trained model.

config_model

List with other arguments that control how the model from Hugging Face is accessed.

Value

A list with the configuration of the model.

Details

A causal language model (also called GPT-like, auto-regressive, or decoder model) is a type of large language model usually used for text-generation that can predict the next word (or more accurately in fact token) based on a preceding context.

If not specified, the causal model that will be used is the one set in specified in the global option pangoling.causal.default, this can be accessed via getOption("pangoling.causal.default") (by default "gpt2"). To change the default option use options(pangoling.causal.default = "newcausalmodel").

A list of possible causal models can be found in Hugging Face website.

Using the config_model and config_tokenizer arguments, it's possible to control how the model and tokenizer from Hugging Face is accessed, see the Python method from_pretrained for details.

In case of errors when a new model is run, check the status of https://status.huggingface.co/

See also

Examples

if (FALSE) { # interactive()
causal_config(model = "gpt2")
}